refactor: moved StorymapTable to components folder
